Move code to fix circular dep
